requirements.txt
appnope==0.1.0
astroid==2.0.4
autopep8==1.4
backcall==0.1.0
bumpversion==0.5.3
decorator==4.3.0
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
lazy-object-proxy==1.3.1
logmatic-python==0.1.7
mccabe==0.6.1
parso==0.3.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycodestyle==2.4.0
Pygments==2.2.0
pylint==2.1.1
python-json-logger==0.1.9
simplegeneric==0.8.1
six==1.11.0
structlog==18.1.0
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
wrapt==1.10.11
